Add shift of FormInputs and renaming with a Form prefix
