+++ src/widgets/FormControl.js
@@ -6,17 +6,17 @@ import { View, ScrollView, StyleSheet } from 'react-native';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { ValidationTextInput } from './ValidationTextInput';
 import { PageButton } from './PageButton';
+import { FormTextInput } from './FormInputs/FromTextInput';
+import { FormDateInput } from './FormInputs/FormDateInput';
+import { FormToggle } from './FormInputs/FormToggle';
+import { FormDropdown } from './FormInputs/FormDropdown';
+import { FormSlider } from './FormInputs/FormSlider';
 
 import globalStyles, { SUSSOL_ORANGE, WHITE } from '../globalStyles';
 import { modalStrings, generalStrings } from '../localization';
 import { selectCanSaveForm, selectCompletedForm } from '../selectors/form';
 import { FormActions } from '../actions/FormActions';
-import { FormDateInput } from './FormDateInput';
-import { FormToggle } from './FormInputs/FormToggle';
-import { FormDropdown } from './FormInputs/FormDropdown';
-import { FormSlider } from './FormInputs/FormSlider';
 
 /**
  * Component which will manage and control a set of user inputs of a form.
@@ -70,7 +70,7 @@ const FormControlComponent = ({
         switch (type) {
           case 'text': {
             return (
-              <ValidationTextInput
+              <FormTextInput
                 ref={refs[index]}
                 onSubmit={nextFocus(index, key)}
                 key={key}
